Fix type changes

+++ solutions/project-euler/001/001.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
+#include <stdint.h>
 
-int16_t main() {
-    int16_t i, sum = 0;
+int main() {
+    int32_t i, sum = 0;
     for (i = 0; i < 1000; i++)
         if (i % 3 == 0 || i % 5 == 0)
             sum += i;
