+++ solutions/project-euler/003/003.c
@@ -1,5 +1,6 @@
 #include <math.h>
 #include <stdio.h>
+#include <stdint.h>
 
 int16_t is_prime(int32_t n) {
     int32_t i;
@@ -19,9 +20,9 @@ int16_t is_prime(int32_t n) {
     return 1;
 }
 
-int16_t main() {
+int main() {
     int16_t i;
-    int32_t n = 600851475143;
+    int64_t n = 600851475143;
     int32_t r;
 
     for (i = 2; i <= sqrt(n); i++) {
