+++ solutions/uri/1026/1026.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
+#include <stdint.h>
 
-int16_t main() {
-    u_int16_t a, b;
+int main() {
+    uint16_t a, b;
 
     while (scanf("%u %u", &a, &b) != EOF) {
         printf("%u\n", a ^ b);
